install the pdfviewer example; provide examples subdirs .pro files
Actually moving the files is unfortunately necessary because of a rule
about having the source directory structure match the intended installation
directory structure.  It was intended to differentiate widget examples from
future QtQuick examples, but now that distiction is being unfortunately lost
in order to keep the directory structure as shallow as it can still be, after
having to comply with the first rule.

Change-Id: I831227d2be3c8f5cab55a98a531e16bcb3aa0303
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>